refactor: downloadAudioFromStorage関数をinfrastructure/storageに移動
