+++ apps/functions/src/infrastructure/firestore/onAudioUpload.ts
@@ -1,5 +1,5 @@
 import { onObjectFinalized } from 'firebase-functions/v2/storage';
-import { downloadAudioFromStorage } from './downloadAudioFromStorage';
+import { downloadAudioFromStorage } from '../storage/downloadAudioFromStorage';
 import { uploadToFileAPI } from '~/lib/gemini/uploadToFileApi';
 import { geminiTranscript } from '~/lib/gemini/geminiTranscript';
 import { saveTranscriptToFirestore } from '~/infrastructure/firestore/saveTranscripts';
